Changes how players are counted for roundstart antags, adds vampire and werewolves (seperately) back to lowpop. (#2227)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Number of players is now determined by the amount of players connected,
not just the ones readied up.
Adds vampire back to lowpop. 50/50 between peasant rebellion and
vampires.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

An unhealthy amount of people don't ready up and play migrants which
affects the roundstart antags. Some purposefully late join to do this as
well.

Adding vampire back to lowpop as an experiment to see how they are
balanced, a rework is planned after bandits are changed.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

---------

Co-authored-by: Gyrandola <pasta.frollagg@gmail.com>