Confession tweaks lite (#2279)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Changes some confession lines and the "Innocent" confession lines
A lite version of this PR
https://github.com/Rotwood-Vale/Ratwood-Keep/pull/2254
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
current atheism lines suck ass and the system is bugged
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->