Barkeep Buffs
The barkeep has been neglected for a long time. They barely have anything that makes them different compared to the rest of the towners. This makes them a bit more fun should trouble start in the inn and the guards aren't there to help.

Why this is good: This lets the barkeep set the cook up well enough and allows them to deal with unruly patrons. Its nothing overwhelmingly strong. Just enough to make the role a bit more interesting.