PR Template Update - Proof of Testing (#2292)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Updates the PR template to include a POT (Proof of testing) section.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

N/A. GitHub stuff.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

Backend change, proof of testing is using this template for this PR.

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->