Letting carpenters do the their job (#2299)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

gives the woodcutter a handsaw
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

woodcutters are supposed to be ALSO carpenters, which the ability to do
so diminished after the pr that added planks and stone bricks as needed
elements for construction.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

It's a single changed line
<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->