makes newmoon spellblades locked to noc (#2307)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

force changes the patron of newmoon spellblades to noc
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

The lore makes them acolytes of noc, this forces them to have them as
their patron, shrimple
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
before selecting newmoon:

![image](https://github.com/user-attachments/assets/83b312bc-b4eb-4feb-92bf-eb4bad453e46)
After:

![image](https://github.com/user-attachments/assets/3f780da1-ebdd-4113-b7b7-8038c7ba49fd)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->