Hoardmaster Salt-PR Gutting (#2110)
## About The Pull Request
Edit: People said that gutting the Hoardmaster would kill bandits,
despite them being fine prior. Ironically, it wasn't much of a problem
to have it entirely REMOVED. Although, they were given lack of an
inspection notice atop that, but it didn't fundamentally break bandits.
Reopened to try this again.
Because of what we saw in testing, this now includes removal of the
'BANDIT!' inspection text.

- - -
Time for the bi-weekly anti-bandit salt PR. Praise be! I know people
will HATE this, but please do read the entire thing and comment WHY you
think this is an objectively bad change.

Effectively removes 90% of Hoardmaster items for classes. Strips it down
to the bare minimum, so it's things they NEED to operate, at best, and a
few sidegrades / upgrades here and there. Bandits will now once again
have to rely on loot, as opposed to favour farming from the zizoid
church smelt strat, the well-deservedly nerfed fishing abuse, and the
myriad other methods to gain favour that I really don't want to repeat
here, otherwise it'll be used for farming, too.

We really ought to make it so you can't open other player's menus, as a
lot of these issues would be solved by that alone, and I wouldn't need
to make a PR like this.

As an aside, this also gives them a light gambeson option to purchase,
cheap as dirt, as a consolation prize.

I originally had this paired with another change, that being the
inclusion of the Forgemaster, within #2023. That wasn't wanted because
of the Forgemaster and so on, as can be see in the eventual PR when I
decoupled it. As a result, this PR is effectively the same, and hasn't
taken into account any further changes from the hoardmaster rebalances
(if any) that have taken part since.

Keep in mind, this doesn't touch ANYTHING regarding Stuff / Things as
categories. They're whatever they were prior, given they weren't touched
in the interim of the initial PR and now this, the exception being the
included gambeson.

## Why It's Good For The Game
Bandits are objectively better equipped by the Hoardmaster alone, once
they've gone through with farming favour, than a majority of the server.
I'll admit that it takes TIME to do so, but it's not difficult. Couple
that with them leveling blacksmithing on a dedicated player? They're
entirely self sufficient and don't need to rely on their namesake.
Pillaging, raiding, etc.
They just have everything handed to them. Add onto that with the fact
that they'll just make the removed equipment anyway through
blacksmithing? AAAAAAAAAA

I really, REALLY dislike that, even as someone who frequently plays
bandit, and as such, decided we needed to change it. As said at the
beginning, I KNOW people will hate this, and it probably won't be
merged, but I feel it appropriate for the health of the server, to at
least SOMEWHAT dial back the capabilities of bandits. I keep getting
told "WAIT ON THE REWORK" but we suffer in the meantime!!!!!!!!!!!

Something I considered was just removing the ability for bandits to do
manual labor crafting, but that might be even worse than this.
Especially since it'd screw stuff like repairs, bridging, the
aforementioned smithing and more.
Another thing was the consideration of making crafted items unable to be
turned in for favour, but that's a majority of what they steal anyways
from something like the Vault, which would be unfair to them
mechanically and OOC faith wise.
I do really think this is the best option, with everything else
considered, though I could just be short sighted in that regard.