hissing! (#2315)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

adds a hissing accent! replacing x with ks and makes s's longer for
lizards and the less refined tabaxi
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

an alternative for the lizard accent that isn't brazillian
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/65f08327-8058-42ce-8a1e-9efae9bd5f89)

![image](https://github.com/user-attachments/assets/25c7f1ea-b86e-44c3-9369-78d2aa7eeebd)

![image](https://github.com/user-attachments/assets/971eba3e-7f25-4579-91e1-c368c00a2c02)

![image](https://github.com/user-attachments/assets/e0c1980f-ed21-41ef-b370-2288c002e13b)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->