Aasimar's wings, customization #2 (#2262)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

- Adds Aasimars some feathery wings.
- Adds new colour skin.


![image](https://github.com/user-attachments/assets/20d94459-afbe-4062-af4c-164164599c26)

- Removes /flight wings var. 
- Removes Angelalt wings from anthro species because they're broken.



![image](https://github.com/user-attachments/assets/75e28ad8-f844-471d-ab2e-5ebb26e59b3c)


## Why It's Good For The Game

More customization for aasimars players because they're not unique and
looks like human but with different colour. This PR helps aasimars be
more special on this build.

Requested by Anya Dawnheart and Twilight Fortress community.

This is second PR because of my skill issue moment with github
https://github.com/Rotwood-Vale/Ratwood-Keep/pull/2233