Arousal freeze + Set arousal option (#2296)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Adds in an arousal freeze and a set arousal button to the ERP menu. The
code was taken from Azure Peak and merged into the sexcon.dm file. I had
planned on just the arousal freeze, but it seems the two are somewhat
connected in the azure code

This was my training attempt to port over code from another server, so i
may learn more and more ss13 coding, and multiple people kept asking for
it so i figured i might as well try for the learning experience. So if
there is ways to make the code better feel free to give me criticism on
it, im doing these to learn

## Why It's Good For The Game

Allows people to freeze their arousal gain during scenes in case they
want to prolong their ERP. People do this through emotes already so now
they can do it mechanically as well to go along with their full
paragraphs they type


## Video Proof of the PR working

https://github.com/user-attachments/assets/24e4ab36-5dc7-45d5-ad51-35f07fe58535


<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->