AIfix and Revivestuff (#2270)
## About The Pull Request

This fixes the AI runtimes caused by mobs trying to flee and having
nowhere to go, this also readds the revival spell to astratan clergy and
disables drawing the divine rune on Useroth's request (Albeit leaves it
in the code for potential use during a later faith rework Useroth has in
mind.

This does _not_ remove the divine rune from the map as we have an active
map testmerge going on, I'm just gonna ping Willy to remove the rune
from the map once this gets merged or TMed

Note: This does not readd the death debuff. That's for a later pr as
part of the faith rework. Also note, this does not fullheal the subject
in question, merely restores them to life and stops bleeding. Ensure
they have enough blood or they will rapidly die again. (Solid way to
stretch the rot timer out a bit at least while you get the corpse into a
viable state)

## Why It's Good For The Game

Useroth wants it, also bugfix.

## Proof Of Testing



https://github.com/user-attachments/assets/23b6e54e-c991-49d8-bf16-6f209e92ddc4

---------

Co-authored-by: Spaghetti-bit <yumyumkillkill@gmail.com>