Barkeep Buffs (#2326)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This is a small buff to the barkeep to let them manage their inn a bit
better. It gives some better stats for both genders, more starting
money, a better wrestling skill, and a buff to their unarmed damage.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
The barkeep is a long neglected role. Other jobs have been given way too
much while this staple was left in the dust. This pull lets them deal
with a rowdy patron on their own and help jumpstart the cook.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->