New rogue subclass! (#2323)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Merges cutthroat and thief, adds a stalker archetype the first refugee
with an actual bow! Other than the tracker, also lets the smuggler
recognize his bottled cargo.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

the cutthroat archetype was pretty bland and nigh exact copy of thief
minus the stealing and sneaking part, trading that in for 1 bonus
strength and medium armor which doesn't fit the class and instead brings
the ranger back adding some more variety.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/e3e3c52c-5164-4384-be58-bed5151767e4)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->