less broken strengthen undead (#2317)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

makes the strengthen undead give a debuff instead of being a stronger
kneestinger stun

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

The strengthen undead spell currently is the most powerful stun spell
that lasts longer than the kneestinger shock while not requiring a
direct line to the target from the caster, being silent and having a
really short cooldown making it pretty borked.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)




https://github.com/user-attachments/assets/3c1cea3c-62a5-41db-a75e-8814171d7a2e




<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->