More dark elf tweaks (#2295)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Continuing iteration of the dark elf light vulnerability based on
playtesting.

- Reworked flash mechanic, now based on difference between last light
update and current light update. Player will not be repeatedly flashed
in bright areas. Cooldown reduced since the trigger condition is now
rarer.
- Increased migraine threshold.
- Reduced potency of migraine debuff.
- Added +1 perception buff if the darkling is in near-total darkness and
has no accumulated light stress. This is pretty rare, and is removed the
second someone walks near them with a torch or lamptern.
- Face covering is now based on whether the face is hidden.
- Adds support for more light resistant garments like sunglasses and
things.
- Removed screen blur effect from eye strain, this was very disliked by
players.
- Face coverings now always block flash effect.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Adjusting based on feedback to make this more interesting, fun to play
with, and less tedious.

Will provide test proof tomorrow.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->



https://github.com/user-attachments/assets/d799a42e-5abf-425c-ad80-5a1d9fe3b396


https://github.com/user-attachments/assets/dcd05969-8773-45b7-a7dc-61cf8f7ef0d4

![image](https://github.com/user-attachments/assets/b42a46f9-9719-4d4a-b007-a5caea5b3dd2)

---------

Co-authored-by: Spaghetti-bit <yumyumkillkill@gmail.com>