Removes wooden planks from the stockpile (#2330)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

This PR removes planks from the stockpile for two main reasons:
1. It prevents people that need to buy wood from doing so, since the
woodcutters simply turn them into planks to win more money. Because for
some reason they're worth more. Leaving the rest of the server without
wood, but with a lot of planks that'll end up unused most of the time,
and just serve the purpose of draining the treasury from money quicker.

2. The ones that do need planks can simply create them themselves by
sawing the wooden log, it's actually cheaper than buying the wooden
planks, so those that DO need the planks don't even buy them.

Should probably also mention issue
https://github.com/Rotwood-Vale/Ratwood-Keep/issues/2253
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Removes an unused material from the stockpile, and stops people from
turning the usable material to a semi-useless one simply for 2 extra
mammons from the stockpile.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Proof of Testing (Required)
Begone planks 

![image](https://github.com/user-attachments/assets/42f9148d-4990-47e0-b088-49f15b963385)
It compiles

![image](https://github.com/user-attachments/assets/a6f789a0-1000-4bea-835b-1f45b5734046)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->