Makes smell messages smaller (#2331)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

As the title says, makes the messages from smells smaller, not only are
they frequent, but they're big, and they can end up cluttering a lot of
the chat logs if you decide to do something like wear perfume. It's
nothing important and mostly flavour, so if the smaller text means it
goes unnoticed by some, it's not that big of a problem.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Smells, something mostly for flavour and almost insignificant, won't be
cluttering the chat with big yellow notice messages, and more important
things will be taking space on the chat.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
Smaller text.

![image](https://github.com/user-attachments/assets/5aa223bc-fda8-4974-9b36-6300c05f7464)
It compiles

![image](https://github.com/user-attachments/assets/b2e412b7-ead3-41e9-bc55-120930ed9c48)


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->