Bandits "uncapped" (#2288)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Changes the bandit cap from a random number between 4 and 8 to a static
8, also changes the bandit to pop ratio to 1:6, meaning it caps out at 8
bandits at 72 player pop removing the random chance to have 4 at high
pop, also brings back bandit migrant waves.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

With all the nerfs the bandits got and the major upgrade to town
security with most important things being moved into town and the
bastion gate being next door to the retinue eager to reinforce them I
think it's fair to bulk the bandit numbers up a bit, since right now
they are not really a major threat, with their low numbers and the lack
of good hoardemaster gear after the recent changes.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->