Firearm Touchup
Arquebus pistol made into a one-handed weapon once again. You can't AIM with it(I THINK? It didn't seem to work in testing.), but you can instantly fire without issue. No accuracy, at the tradeoff of instant response. Returning to how it used to be.

Adjusts the ammo pouch size so that it can be shoved into bags and the such, as the same size as pouches.

Atop that, it gives the Inquisitor their proper runelock pouch, removing the old three-round pouch.