Harlequin Markings for the Furries + Lynx tail fix (#2337)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Furries Rejoice . Harlequin markings.

Do you wanna be a literal jester that's also a harlequin? JOY FOR YOU.

Harlequin markings (body and head) will ALLOW YOU to be a freak
(assuming you're a Tabaxi, Lupian, Vulpkian or Wildkin). I didn't add
the limbs because... plain exists.

Just use plain for limbs. Dude.

Also fixed the lynx tail that was just. Nine pixels, literally. I
couldn't be arsed. I made it a bit more prominent. And no, no wag yet.
Or ever. Who knows.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Customization that will hopefully not be used for the 9th sparkledog
prostitute in the week.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

# BEWARE OF TITTIES....

![image](https://github.com/user-attachments/assets/86439bbc-925b-4204-adb4-28a780a61469)

![image](https://github.com/user-attachments/assets/3511b7da-6af4-4aa4-af5d-cab99efb6a03)

Male and female, both sides of the harlequin

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->