Smoke bombs! (#2334)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Adds craftable smoke bombs that explode in a 3x3 square of opaque smoke,
useful tool for all rogue-ish types, also gives said bombs to some
rogue-ish types

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Another interesting tool for people to use to get in and out of combat
is nice. Also it's a long requested feature by a good part of the
community
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


https://github.com/user-attachments/assets/5b1fe2b6-2369-4cfe-8ded-fbca16444e59


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->