Fixes the belt longsword sprites (#2358)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Longsword sprites were bugged, and most of them looked like this:

![image](https://github.com/user-attachments/assets/b0811902-49a1-4c24-973f-63b984947d85)

This fixes the problem, and makes them look properly on the belts:
In order, these are: Bastard sword, old sword, crimson fang (vlord),
judgement (unobtainable unless admin spawned) and execution sword.

![image](https://github.com/user-attachments/assets/c557bc2b-b442-4e0f-8872-4f4e70864b4a)
![image](https://github.com/user-attachments/assets/564bfe69-c604-4dfd-85fe-be6785a649f6)
![image](https://github.com/user-attachments/assets/dda8f5d9-e392-4fec-bf43-787cfccfcafc)
![image](https://github.com/user-attachments/assets/258c73d2-e54b-49b8-98ff-3c19b5c8a4b8)
![image](https://github.com/user-attachments/assets/178dcfcc-c42b-4145-ad2d-1955e3b1aa2e)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Bug fix! And it also lets you see that your enemy does, in fact, carry a
big ass word on his hip. You also no longer have to reach down on wierd
angles to grab the handle of the sword, you're welcome.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/98b6fb3f-af8e-485e-bd5a-9e8c43354d36)

And also all of the photos i put up on the about section.
<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->