Adds Marking Customizations to Sissean + Kobold (#2356)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Ports over markings that make sense from Wildkin/Verminvolk over to
Sissean/Kobolds for their respective sizes. Such as verminvolk's
markings to kobolds, wildkin to sissean. Does not touch customization
features only the actual markings themselves.

## Why It's Good For The Game

As someone who has made a ton of characters it is really dumb how we
lack markings on certain races to get that multihued/shaded look. I am a
person who likes detailed characters and having more customization
options lets players make more unique characters. Such as a brown kobold
with gold spots. Or a sissean that looks like a gecko. Sure, make it go
to wildkin/verminvolk. But then you don't get their language, or you get
to be a wannabe of their race. This is a pretty harmless PR but a needed
one in my opinion.

## Proof of Testing (Required)

![siss](https://github.com/user-attachments/assets/2ba248c1-11a4-404f-abae-63ecfb67d0eb)

![kobo](https://github.com/user-attachments/assets/b609ad5e-2623-4eed-961b-685303b313c3)

![exampleko](https://github.com/user-attachments/assets/3c5f21fe-4caa-4f29-bb18-7b2e9467d551)

![stupidko](https://github.com/user-attachments/assets/fd61d7f4-b9b3-4acb-b351-d45665821a18)