Sneaky councillors (#2366)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Makes Councillors more akin to Spymasters than brawlers?
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

There is no real reason for them to have those skills as fragile nobles,
and courtiers which would more so benefit from guile and intrigue than
brawn, and they get a dagger.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/924b8039-b9e5-46e6-bc39-492f80ccc829)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->