Re adds Engineering 2 to blacksmith towner, upd tutorial, reverts blue dress colour (#2352)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Updates the tutorial to better reflect the role, re adds engineering 2
that was removed by mistake(?)
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Fixes an error, tutorial was pissing me off too.
https://github.com/wow50000/Ratwood-Keep/commit/735131b60b533830e496dcbd2400c013e9001b63#diff-92f0fa9eecde8da1c15118a037ad877f4eaf47dc0bb1c9e1d29b692ab5bcd564
removed in this commit

![image](https://github.com/user-attachments/assets/9304488f-63c0-4c40-82d0-c55b9967111e)

![image](https://github.com/user-attachments/assets/d60dd9d7-c21e-4385-8800-4b8ef552f795)

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
you serious?
<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->