Adds ponytail 8 and fennec ears to vulp (#2384)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Adds ponytail 8 and fennec ears as options for Vulps. Requested on
Discord by Fulpie, and I am susceptible to bribery.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

So fox players can actually be foxes with their fox people. Also because
I was offered a sprite token to make the PR.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


![image](https://github.com/user-attachments/assets/a049d781-95fc-455b-abb9-a3a2195f49b9)



<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->