Making planks now correctly uses carpentry (#2333)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

For some reason, cutting planks needed carpentry to do, and it gave you
experience on lumberjacking. Now it gives you experience and uses
lumberjacking correctly.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Bug fix!
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


![image](https://github.com/user-attachments/assets/5f234930-3f12-47c9-89e3-88af5ca07223)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->