Zybantine combat music (#2347)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Adds zybantine combat music as requested:
https://github.com/Rotwood-Vale/Ratwood-Keep/issues/2341 and extends it
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

combat music was forgotten about in the zybantine update, it was already
in the files.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/1b28afaf-a3dc-47f9-afb1-17736501c636)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->