All natural tables (#2346)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Makes tree stumps work as tables, shrimple (probably did it the worst
way possible but it works)
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Both tree stumps and tables are big flat surfaces and there really isn't
a reason you shouldn't be able to use them as you would use a table,
also lets hunters make bait without having to run to town or ruins.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


https://github.com/user-attachments/assets/f40659f1-8651-4441-9e08-28ade4dd9251


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->