Premium slaves and exotic silks (#2345)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

steals exotic silks and anklets from dreamkeep and gives them to the
zybantine slave and for the merchant to buy, also adds a bit of code to
make sole-less shoes work
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
more variety and identity for regions, also more end-game stuff to spend
your money on, can't go wrong with either, also the zybantine slaves are
virgins, thus premium cargo, it makes zero sense for them to be dressed
in shitty rags.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/5724877f-e583-44dc-bf44-3bd4713dc7e2)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->